+++ src/shared/ui/styles/footer/footer.css.ts
