+++ src/shared/ui/styles/sidebar/sidebar.css.ts
